requirementst.txt
streamlit
python-docx
gtts
deep-translator
SpeechRecognition
nltk


